+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Cinecred Changelog
   centered using a new spine position option.
 - Added more fine-grained controls for vertical head and tail justification.
 - Added a dedicated coloring option for hiding helper layers.
+- Added the option to specify a layer's offset in polar coordinates.
 
 
 1.4.1